refactored clean architecture
